fix conversion error (invalid narrowing conversion from "int" to "float")

+++ main.cpp
@@ -221,7 +221,7 @@ int main()
         }
 
         ClearBackground(green);
-        DrawRectangleLinesEx(Rectangle{offset - 5, offset - 5, cellSize * cellCount + 10, cellSize * cellCount + 10}, 5, darkGreen);
+        DrawRectangleLinesEx(Rectangle{(float)offset - 5, (float)offset - 5, (float)cellSize * cellCount + 10, (float)cellSize * cellCount + 10}, 5, darkGreen);
         game.Draw();
 
         EndDrawing();
